recordstream/parser/pyproject.toml
[tool.poetry]
name = "hedera"
version = "0.1.0"
description = "Opensource recordstream"
authors = ["Ngan <ngan88le@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
numpy = "^1.21.0"
pandas = "^1.3.1"
jsonlines = "2.0.0"
google = "3.0.0"
google-cloud = "0.34.0"
google-cloud-storage = "1.38.0"
psutil = "5.8.0"
pydantic = "^1.9.0"
protobuf = "<4.0.0"
python-dotenv = "0.17.1"
pretty-html-table = "0.9.7"
ray = "2.5.1"
pendulum = "^2.1.2"
typer = "^0.6.1"
black = "^22.10.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"
pylint = "^2.15.8"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
requests-mock = "^1.10.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
flake8 = "^6.0.0"
flake8-bugbear = "23.3.23"
autopep8 = "^1.5.4"

[tool.poetry.scripts]
record-file-downloader = 'hedera.cli:record_file_downloader'
record-file-orchestrator = 'hedera.cli:record_file_orchestrator'
saucerswap-orchestrator = 'hedera.cli:saucerswap_orchestrator'
hedera-app = 'hedera.cli:app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
